SpinWaveTheory

Introduction

Spin wave theory for magnetically ordered quantum lattice systems based on the QuantumLattices pack and the TightBindingApproximation pack.

Installation

In Julia v1.6+, please type ] in the REPL to use the package mode, then type this command:

pkg> add SpinWaveTheory

Getting Started

Examples of spin wave theory for magnetically ordered quantum lattice systems

Manuals

QuantumLattices.DegreesOfFreedom.Hilbert — Method

Hilbert(hilbert::Hilbert{<:Spin}, magneticstructure::MagneticStructure)

Get the corresponding Hilbert space of the original one after the Holstein-Primakoff transformation.

QuantumLattices.DegreesOfFreedom.Metric — Method

Metric(::Magnonic, hilbert::Hilbert{<:Fock{:b}}) -> OperatorUnitToTuple

Get the index-to-tuple metric for a quantum spin system after the Holstein-Primakoff transformation.

SpinWaveTheory.HPTransformation — Type

HPTransformation{S<:Operators, U<:CompositeIndex, M<:MagneticStructure} <: AbstractUnitSubstitution{U, S}

Holstein-Primakoff transformation.

SpinWaveTheory.LSWT — Type

LSWT{K<:TBAKind{:BdG}, L<:AbstractLattice, H<:OperatorGenerator, HP<:HPTransformation, E<:Image, F<:Image} <: AbstractTBA{K, H, AbstractMatrix}

Linear spin wave theory for magnetically ordered quantum lattice systems.

SpinWaveTheory.LSWT — Method

LSWT(lattice::AbstractLattice, hilbert::Hilbert{<:Spin}, terms::Tuple{Vararg{Term}}, magneticstructure::MagneticStructure; neighbors::Union{Nothing, Int, Neighbors}=nothing, boundary::Boundary=plain)

Construct a LSWT.

SpinWaveTheory.MagneticStructure — Type

MagneticStructure{L<:AbstractLattice, D<:Number}

The magnetic structure of an ordered quantum lattice system.

SpinWaveTheory.MagneticStructure — Method

MagneticStructure(cell::AbstractLattice, moments::Dict{Int, <:AbstractVector})

Construct the magnetic structure on a given lattice with the given moments.

SpinWaveTheory.Magnonic — Type

Magnonic <: TBAKind{:BdG}

Magnonic quantum lattice system.

TightBindingApproximation.TBAMatrixRepresentation — Type

TBAMatrixRepresentation(lswt::LSWT, k=nothing, gauge::Symbol=:icoordinate)

Construct the matrix representation transformation of a quantum spin system using the linear spin wave theory.

QuantumLattices.QuantumOperators.matrix — Method

matrix(lswt::LSWT; k=nothing, gauge=:icoordinate, atol=atol/5, kwargs...) -> TBAMatrix

Get the tight-binding matrix representation of the linear spin waves.

Here, the atol parameter is used to ensure that the matrix is positive-definite so that the Cholesky decomposition can be performed numerically.

QuantumLattices.add! — Method

add!(dest::Matrix, mr::TBAMatrixRepresentation{<:LSWT}, m::Operator{<:Number, <:ID{CompositeIndex{<:Index{Int, <:FID{:b}}}}}; atol=atol/5, kwargs...) -> typeof(dest)

Get the matrix representation of an operator and add it to destination.

SpinWaveTheory.rotation — Method

rotation(destination::AbstractVector{<:Number}) -> SMatrix{3, 3}

Get the rotation matrix which rotates the [0, 0, 1] or [θ, ϕ] vector to the direction of the destination vector.

TightBindingApproximation.commutator — Method

commutator(::Magnonic, hilbert::Hilbert{<:Fock{:b}}) -> Diagonal

Get the commutation relation of the Holstein-Primakoff bosons.