SpinWaveTheory

Introduction

Spin wave theory for magnetically ordered quantum lattice systems based on the QuantumLattices pack and the TightBindingApproximation pack.

Installation

In Julia v1.6+, please type ] in the REPL to use the package mode, then type this command:

pkg> add SpinWaveTheory

Getting Started

Examples of spin wave theory for magnetically ordered quantum lattice systems

Manuals

QuantumLattices.Essentials.DegreesOfFreedom.Hilbert — Method

Hilbert(hilbert::Hilbert{<:Spin}, magneticstructure::MagneticStructure)

Get the corresponding Hilbert space of the original one after the Holstein-Primakoff transformation.

SpinWaveTheory.HPTransformation — Type

HPTransformation{S<:Operators, U<:OID{<:Index{<:AbstractPID, <:SimpleIID}}, M<:MagneticStructure} <: AbstractUnitSubstitution{U, S}

Holstein-Primakoff transformation.

SpinWaveTheory.InelasticNeutronSpectra — Type

InelasticNeutronSpectra{P<:ReciprocalPath, E<:AbstractVector} <: Action

Inelastic neutron spectra of magnetically ordered quantum lattice systems by linear spin wave theory.

SpinWaveTheory.LSWT — Type

LSWT{L<:Lattice, H<:Generator, HP<:HPTransformation, E<:Image, F<:Image} <: AbstractTBA{TBAKind(:BdG), H, AbstractMatrix}

Linear spin wave theory for magnetically ordered quantum lattice systems.

SpinWaveTheory.LSWT — Method

LSWT(lattice::Lattice, hilbert::Hilbert, terms::Tuple{Vararg{Term}}, magneticstructure::MagneticStructure; boundary::Boundary=plain)

Construct a LSWT.

SpinWaveTheory.MagneticStructure — Type

MagneticStructure{L<:Lattice, P<:AbstractPID, D<:Number}

The magnetic structure of an ordered quantum lattice system.

SpinWaveTheory.MagneticStructure — Method

MagneticStructure(cell::Lattice, moments::Dict{<:AbstractPID, <:AbstractVector})

Construct the magnetic structure on a given lattice with the given moments.

TightBindingApproximation.TBAMatrixRepresentation — Type

TBAMatrixRepresentation(lswt::LSWT, k=nothing, gauge::Symbol=:icoord)

Construct the matrix representation transformation of a quantum spin system using the linear spin wave theory.

QuantumLattices.Essentials.QuantumOperators.matrix — Method

matrix(lswt::LSWT; k=nothing, gauge=:icoord, atol=atol/5, kwargs...) -> TBAMatrix

Get the tight-binding matrix representation of the linear spin waves.

Here, the atol parameter is used to ensure that the matrix is positive-definite so that the Cholesky decomposition can be performed numerically.

QuantumLattices.Interfaces.add! — Method

add!(dest::Matrix,
    mr::TBAMatrixRepresentation{<:LSWT},
    m::Operator{<:Number, <:ID{CompositeOID{<:Index{<:AbstractPID, <:FID{:b}}}}};
    atol=atol/5,
    kwargs...
    ) -> typeof(dest)

Get the matrix representation of an operator and add it to destination.

SpinWaveTheory.rotation — Method

rotation(destination::AbstractVector{<:Number}) -> SMatrix{3, 3}

Get the rotation matrix which rotates the [0, 0, 1] vector to the direction of the destination vector.